requirements.txt
docutils==0.16
Flask==1.1.2
imageio==2.8.0
imagesize==1.2.0
img2pdf==0.3.6
imutils==0.5.3
Jinja2==2.11.2
numpy==1.16.2
opencv-python==4.2.0.34
pandas==0.24.2
Pillow==5.4.1
requests==2.24.0
Werkzeug==1.0.1
future==0.18.2
gunicorn==20.0.4
